Move files to dir
