+++ core-ui/src/components/Predefined/Details/Secret/helm/ChartContent.js
