+++ core-ui/src/components/Predefined/Details/Secret/helm/HelmReleaseData.js
