+++ core-ui/src/components/Predefined/Details/Secret/helm/ReleaseData.js
